fix: incorporate separate of concerns, added comments
