+++ src/ui/common/components/Activity/components/ActivityList.tsx
@@ -2,161 +2,26 @@ import { useMemo } from "react";
 
 import { ExpansionHistoryModal } from "@/ui/common/components/ExpansionHistory/ExpansionHistoryModal";
 import { getNetworkConfig } from "@/ui/common/config/network";
-import { DELEGATION_ACTIONS as ACTIONS } from "@/ui/common/constants";
-import {
-  ActionType,
-  useDelegationService,
-} from "@/ui/common/hooks/services/useDelegationService";
+import { useDelegationService } from "@/ui/common/hooks/services/useDelegationService";
 import { useStakingManagerService } from "@/ui/common/hooks/services/useStakingManagerService";
 import { useFinalityProviderState } from "@/ui/common/state/FinalityProviderState";
 import { useStakingExpansionState } from "@/ui/common/state/StakingExpansionState";
 import {
+  DelegationV2,
   DelegationV2StakingState,
   DelegationWithFP,
 } from "@/ui/common/types/delegationsV2";
-import { FinalityProviderState } from "@/ui/common/types/finalityProviders";
+
+import { ActivityCard } from "../../ActivityCard/ActivityCard";
+import { getActionButton } from "../../ActivityCard/utils/actionButtonUtils";
 import {
   ActivityCardTransformOptions,
   transformDelegationToActivityCard,
-} from "@/ui/common/utils/activityCardTransformers";
-
-import {
-  ActivityCard,
-  ActivityCardActionButton,
-} from "../../ActivityCard/ActivityCard";
+} from "../../ActivityCard/utils/activityCardTransformers";
 import { DelegationModal } from "../../Delegations/DelegationList/components/DelegationModal";
 import { StakingExpansionModalSystem } from "../../StakingExpansion/StakingExpansionModalSystem";
 
 const networkConfig = getNetworkConfig();
-
-const getActionButton = (
-  delegation: DelegationWithFP,
-  onAction: (action: ActionType, delegation: DelegationWithFP) => void,
-  isStakingManagerReady: boolean,
-): ActivityCardActionButton | undefined => {
-  const { state, fp } = delegation;
-
-  // Define action mapping
-  const actionMap: Record<
-    string,
-    Record<string, { action: ActionType; title: string }>
-  > = {
-    [FinalityProviderState.ACTIVE]: {
-      [DelegationV2StakingState.VERIFIED]: {
-        action: ACTIONS.STAKE,
-        title: "Stake",
-      },
-      [DelegationV2StakingState.ACTIVE]: {
-        action: ACTIONS.UNBOND,
-        title: "Unbond",
-      },
-      [DelegationV2StakingState.EARLY_UNBONDING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_EARLY_UNBONDING,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.TIMELOCK_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_TIMELOCK,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.TIMELOCK_SLASHING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_TIMELOCK_SLASHING,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.EARLY_UNBONDING_SLASHING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_EARLY_UNBONDING_SLASHING,
-        title: "Withdraw",
-      },
-    },
-    [FinalityProviderState.INACTIVE]: {
-      [DelegationV2StakingState.VERIFIED]: {
-        action: ACTIONS.STAKE,
-        title: "Stake",
-      },
-      [DelegationV2StakingState.ACTIVE]: {
-        action: ACTIONS.UNBOND,
-        title: "Unbond",
-      },
-      [DelegationV2StakingState.EARLY_UNBONDING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_EARLY_UNBONDING,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.TIMELOCK_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_TIMELOCK,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.TIMELOCK_SLASHING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_TIMELOCK_SLASHING,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.EARLY_UNBONDING_SLASHING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_EARLY_UNBONDING_SLASHING,
-        title: "Withdraw",
-      },
-    },
-    [FinalityProviderState.JAILED]: {
-      [DelegationV2StakingState.VERIFIED]: {
-        action: ACTIONS.STAKE,
-        title: "Stake",
-      },
-      [DelegationV2StakingState.ACTIVE]: {
-        action: ACTIONS.UNBOND,
-        title: "Unbond",
-      },
-      [DelegationV2StakingState.EARLY_UNBONDING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_EARLY_UNBONDING,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.TIMELOCK_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_TIMELOCK,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.TIMELOCK_SLASHING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_TIMELOCK_SLASHING,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.EARLY_UNBONDING_SLASHING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_EARLY_UNBONDING_SLASHING,
-        title: "Withdraw",
-      },
-    },
-    [FinalityProviderState.SLASHED]: {
-      [DelegationV2StakingState.ACTIVE]: {
-        action: ACTIONS.UNBOND,
-        title: "Unbond",
-      },
-      [DelegationV2StakingState.EARLY_UNBONDING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_EARLY_UNBONDING,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.TIMELOCK_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_TIMELOCK,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.TIMELOCK_SLASHING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_TIMELOCK_SLASHING,
-        title: "Withdraw",
-      },
-      [DelegationV2StakingState.EARLY_UNBONDING_SLASHING_WITHDRAWABLE]: {
-        action: ACTIONS.WITHDRAW_ON_EARLY_UNBONDING_SLASHING,
-        title: "Withdraw",
-      },
-    },
-  };
-
-  const actionConfig = actionMap[fp?.state]?.[state];
-  if (!actionConfig) return undefined;
-
-  const isUnbondDisabled =
-    state === DelegationV2StakingState.ACTIVE && !isStakingManagerReady;
-
-  return {
-    label: actionConfig.title,
-    onClick: () => onAction(actionConfig.action, delegation),
-    variant: "contained",
-    size: "medium",
-    className: isUnbondDisabled ? "opacity-50" : "",
-  };
-};
 
 export function ActivityList() {
   const {
@@ -194,17 +59,43 @@ export function ActivityList() {
       )
       .map((delegation) => {
         const options: ActivityCardTransformOptions = {
-          showActions: true,
           showExpansionSection: true,
-          onAction: openConfirmationModal,
-          isStakingManagerReady,
-          getActionButton,
         };
-        return transformDelegationToActivityCard(
+        const cardData = transformDelegationToActivityCard(
           delegation,
           finalityProviderMap,
           options,
         );
+
+        // Create delegation with FP for action button
+        // delegations from useDelegationService are DelegationV2 objects
+        const delegation_v2 = delegation as DelegationV2;
+        // Use the first FP [0] for backward compatibility with action button logic
+        // which expects a single FP to determine button state. The full BSN/FP pairs
+        // are properly displayed in the card's grouped details section
+        const fp =
+          Array.isArray(delegation_v2.finalityProviderBtcPksHex) &&
+          delegation_v2.finalityProviderBtcPksHex.length > 0
+            ? finalityProviderMap.get(
+                delegation_v2.finalityProviderBtcPksHex[0],
+              )
+            : undefined;
+        const delegationWithFP: DelegationWithFP = {
+          ...delegation_v2,
+          fp,
+        } as DelegationWithFP;
+
+        // Add action button if applicable
+        const primaryAction = getActionButton(
+          delegationWithFP,
+          openConfirmationModal,
+          isStakingManagerReady,
+        );
+
+        return {
+          ...cardData,
+          primaryAction,
+        };
       });
   }, [
     delegations,
