+++ src/ui/common/hooks/useExpansionHistoryModalData.ts
@@ -1,12 +1,12 @@
 import { useMemo } from "react";
 
+import { transformDelegationToActivityCard } from "@/ui/common/components/ActivityCard/utils/activityCardTransformers";
 import { useExpansionHistoryService } from "@/ui/common/hooks/services/useExpansionHistoryService";
 import { useFinalityProviderState } from "@/ui/common/state/FinalityProviderState";
 import {
   DelegationV2,
   DelegationWithFP,
 } from "@/ui/common/types/delegationsV2";
-import { transformDelegationToActivityCard } from "@/ui/common/utils/activityCardTransformers";
 
 import { ActivityCardData } from "../components/ActivityCard/ActivityCard";
 
@@ -39,9 +39,8 @@ export function useExpansionHistoryModalData({
 
     return historyChain.map((delegation, index) => {
       const stepLabel = index === 0 ? "Original Stake" : `Expansion ${index}`;
-      // Options for expansion history: no actions, no expansion section, hide expansion completely
+      // Options for expansion history: no expansion section, hide expansion completely
       const options = {
-        showActions: false,
         showExpansionSection: false,
         hideExpansionCompletely: true,
       };
